Consistently use argparse for argument parsing
This change updates all of the commands to consistently use "argparse"
for argument parsing; and the "Command" class has been modified to
enforce this.

Now, when a "Command" object is created, it will create an argument
parser that doesn't have any supported arguments. If a specific command
wishes to support arguments, that command's class must implement the
"_init_argparse" method, to add its supported arguments to the parser.